moved memory usage to its own file
