+++ iGlance/iGlance/Components/CpuUsageComponent.swift
@@ -9,7 +9,6 @@
 import Cocoa
 
 class CpuUsageComponent {
-    
     // the status item of the cpu utilization
     static let sItemCpuUtil = NSStatusBar.system.statusItem(withLength: 27.0)
     // the custom menu view of the cpu utilization
@@ -20,7 +19,7 @@ class CpuUsageComponent {
     var btnCpuUtil: NSStatusBarButton?
     // the menu for the button
     var menuCpuUtil: NSMenu?
-    
+
     /**
      * CPU Button Image variables
      */
@@ -33,10 +32,10 @@ class CpuUsageComponent {
     var pixelWidth: Double?
     var pbImg: String?
     var pbMax: Double?
-    
+
     // system variable to get cpu stats
     var mySystem: System?
-    
+
     init() {
         pbMax = 16.0 // 32*0.5
         pixelWidth = 7 // 14*0.5
@@ -45,10 +44,10 @@ class CpuUsageComponent {
         btnCpuUtil = CpuUsageComponent.sItemCpuUtil.button
         btnCpuUtil?.image?.isTemplate = true
     }
-    
+
     func initialize() {
         menuItemCpu.view = myCpuMenuView
-        
+
         menuCpuUtil = NSMenu()
         menuCpuUtil?.addItem(menuItemCpu)
         menuCpuUtil?.addItem(NSMenuItem.separator())
@@ -57,20 +56,20 @@ class CpuUsageComponent {
         menuCpuUtil?.addItem(NSMenuItem(title: "Quit iGlance", action: #selector(NSApplication.terminate(_:)), keyEquivalent: "q"))
         CpuUsageComponent.sItemCpuUtil.menu = menuCpuUtil
     }
-    
+
     func updateCPUUsage() {
         let cpuStats = mySystem!.usageCPU()
         let cpuUser = Double(round(100 * cpuStats.user) / 100)
         let cpuSystem = Double(round(100 * cpuStats.system) / 100)
         let cpuIdle = Double(round(100 * cpuStats.idle) / 100)
         let cpuUsageTotal = cpuUser + cpuSystem
-        
+
         myCpuMenuView.percentSystem.stringValue = String(Int(cpuSystem)) + "%"
         myCpuMenuView.percentUser.stringValue = String(Int(cpuUser)) + "%"
         myCpuMenuView.percentIdle.stringValue = String(Int(cpuIdle)) + "%"
-        
+
         pixelHeightCpu = Double((pbMax! / 100.0) * cpuUsageTotal)
-        
+
         if InterfaceStyle() == InterfaceStyle.Dark {
             cpuImg = "menubar-label-cpu-white"
             pbImg = "progressbar-white"
@@ -91,7 +90,7 @@ class CpuUsageComponent {
         pbFillRectCpu?.fill()
         NSColor.clear.setFill()
         imgFinal.unlockFocus()
-        
+
         btnCpuUtil?.image = imgFinal
     }
 }
